--- a/package.json
+++ b/package.json
@@ -8,7 +8,9 @@
     "dist"
   ],
   "scripts": {
-    "test": "jest --config config/jest.config.ts",
+    "test": "npm run test:unit",
+    "test:unit": "jest --config config/jest.unit.config.ts",
+    "test:performance": "jest --config config/jest.performance.config.ts",
     "test:coverage": "jest --config config/jest.config.ts --coverage",
     "build": "tsc -p config/tsconfig.json"
   },