… tests/performance/cases/geofences-case-1.json
… tests/performance/cases/geofences-case-2.json
… tests/performance/cases/geofences-case-3.json
… tests/performance/cases/geofences-case-4.json
… tests/performance/cases/geofences-case-5.json
… tests/performance/cases/geofences-case-6.json
… tests/unit/geofences.json