This PR refactors the geofence reducer library by replacing the substring-based uniqueness check with a frequency-based algorithm, improving the efficiency of geohash processing while maintaining input-output consistency.  

#### Summary of Results

| Precision | Time (Original) | Time (Alternative) | Speedup Factor | Number of Geofences |
|-----------|-----------------|--------------------|----------------|----------------------|
| 6         | 569,936.63 ms  | 1,212.32 ms       | ~470.12x       | 2,226               |
| 5         | 20,601.51 ms   | 49.96 ms          | ~412.33x       | 2,028               |
| 4         | 2,002.47 ms    | 6.13 ms           | ~326.91x       | 2,118               |

For additional details and extended results, see the [Geofence Reducer Library Improvement document](https://github.com/user-attachments/files/18440205/Public.-.Geofence.Reducer.Library.Improvement.pdf).