@filipecorrea 2024-12-18
Hi @InfoLorenzo. Can you share the performance tests and results in this PR description?