@InfoLorenzo 2024-12-23
> Hi @InfoLorenzo. Can you share the performance tests and results in this PR description?

Hi @filipecorrea, It’s great to hear from you again! I’m currently on holiday, but I’ll post updates as soon as I’m back.
Thanks for getting involved, I really appreciate it!