@InfoLorenzo 2025-01-13
Hey @filipecorrea,

While testing, I noticed that the same geohash sometimes appears multiple times for a single geofence, which is causing discrepancies in my frequency map. I also checked the vicinityhash library and saw the same geohash show up four times for a single geofence at precision 5.

Which is the reason behind why we're getting these duplicates when “compress” is false? Could it be related to how we mirror coordinates in different quadrants, or is there another reason?

I would love to hear your thoughts on this. Thank you in advance!