@InfoLorenzo 2025-01-16
Hi @filipecorrea,

Just a quick update!
- Added the performance tests. There’s also a [document](https://docs.google.com/document/d/e/2PACX-1vSq29i7qhH4hG0SITzvspFJ4yf42oXgtuCITSHzCLOM-4YRAYRp_XinmPfpnjLnUMwiFSFVbzz1F6Ly/pub) with the results to explain the changes and shows some results.
- Updated the unit test to check that the input and output geohashes cover the same areas with the same precision. This makes sure everything is correct.
- Added a uniquify step to handle the duplicate geohashes from vicinityhash for now, ensuring the results remain consistent.

Let me know what you think when you have time! 😊