@InfoLorenzo 2025-01-16
> Maybe there's an issue with [vicinityhash](https://github.com/filipecorrea/vicinityhash) compression part, because it's intent is to remove the geohashes duplicates and use a single, broader, geohash to represent the same area. Can you take a look and see if you can find the issue? I think no one reviewed this code so far and I really appreciate if you could do so.

Sure! I did some testing, and the compression works quite well overall. The only issue is that when compression is set to false, the library can sometimes return the same geohash a few times for the same geofence. This happens because we don’t use a Set to handle the geohashes in this case, unlike when compression is set to true.

One potential fix could be to use a Set even when compression is false. However, based on my tests, this approach caused a 30% performance drop, significantly slowing down execution. Another option would be to explore the rest of the code for a more efficient solution.

> I can't access your document, unfortunately, but I think I got the picture by reading the code change.

Ah, I see. I thought the document was accessible, but it seems it’s not. I’ve attached a PDF to this comment. The design isn’t great, but all the data is included.

[Public - Geofence Reducer Library Improvement.pdf](https://github.com/user-attachments/files/18440205/Public.-.Geofence.Reducer.Library.Improvement.pdf)

> I can't see your updates on the unit tests though. Are you sure you pushed it?

Just pushed them now! I had a bit of trouble with the GPG keys, classic development hiccup, haha.