@InfoLorenzo 2025-01-16
In case you would like to see from where I've got those numbers, here is the full report:

[Performance Comparison (precision 4..6).pdf](https://github.com/user-attachments/files/18440322/Performance.Comparison.precision.4.6.pdf)