@InfoLorenzo 2025-01-17
> Thanks for sending the report. You made my day!
> 
> You did a great job and the changes are approved from my side, but I don't have push access to this repo anymore. Can you merge it?

You made my day too! Thank you for the quick review and for getting involved on this 💪

I held off on merging earlier since I wasn’t able to, but now that I can, I’ll go ahead and do it. Thank you @MartinodF!